Refactoring assets crud
