+++ flexmeasures/ui/crud/assets/__init__.py
@@ -1,8 +1,10 @@
-from flexmeasures.ui.crud.assets.assets import (
-    AssetCrudUI,
+from flexmeasures.ui.crud.assets.forms import AssetForm, NewAssetForm
+from flexmeasures.ui.crud.assets.utils import (
+    get_allowed_price_sensor_data,
+    get_allowed_inflexible_sensor_data,
     process_internal_api_response,
     user_can_create_assets,
     user_can_delete,
     get_assets_by_account,
 )
-from flexmeasures.ui.crud.assets.services import AssetForm, NewAssetForm
+from flexmeasures.ui.crud.assets.views import AssetCrudUI
